+++ demisto_sdk/commands/validate/validators/BA_validators/BA106_is_from_version_sufficient_integration.py
@@ -27,8 +27,6 @@ class IsFromVersionSufficientIntegrationValidator(
 ):
     description = "Validate that an integration has a high enough fromversion field according to whether it's a powershell/feed/regular."
     error_message = "The integration is a {0} integration and therefore require a fromversion field of at least {1}, current version is: {2}."
-    related_field = "fromversion"
-    is_auto_fixable = True
 
     def is_valid(self, content_items: Iterable[ContentTypes]) -> List[ValidationResult]:
         return [
